+++ source/git-api.js
@@ -8,11 +8,10 @@ const rimraf = require('rimraf');
 const _ = require('lodash');
 const gitPromise = require('./git-promise');
 const fs = require('./utils/fs-async');
+const watch = require('node-watch');
 const ignore = require('ignore');
 const Bluebird = require('bluebird');
 
-const isMac = /^darwin/.test(process.platform);
-const isWindows = /^win/.test(process.platform);
 const tenMinTimeoutMs = 10 * 60 * 1000;
 
 exports.pathPrefix = '';
@@ -41,16 +40,6 @@ exports.registerApi = (env) => {
           .then(() => {
             socket.watcher = [];
             return watchPath(socket, '.', {'recursive': true});
-          }).then(() => {
-            if (!isMac && !isWindows) {
-              // recursive fs.watch only works on mac and windows
-              const promises = [];
-              promises.push(watchPath(socket, path.join('.git', 'HEAD')));
-              promises.push(watchPath(socket, path.join('.git', 'refs', 'heads')));
-              promises.push(watchPath(socket, path.join('.git', 'refs', 'remotes')));
-              promises.push(watchPath(socket, path.join('.git', 'refs', 'tags')));
-              return Bluebird.all(promises);
-            }
           }).finally(callback);
       });
     });
@@ -66,10 +55,10 @@ exports.registerApi = (env) => {
           return mkdirp(pathToWatch);
         }
       }).then(() => {
-        const watcher = fs.watch(pathToWatch, options || {});
-        watcher.on('change', (event, filename) => {
-          if (!filename) return;
-          const filePath = path.join(subfolderPath, filename);
+        const watcher = watch(pathToWatch, options || {});
+        watcher.on('change', (event, changedPath) => {
+          if (!changedPath) return;
+          const filePath = path.relative(socket.watcherPath, changedPath);
           winston.debug(`File change: ${filePath}`);
           if (isFileWatched(filePath, socket.ignore)) {
             winston.info(`${filePath} triggered refresh for ${socket.watcherPath}`);